make the build status fail if test fails

+++ test.sh
@@ -6,4 +6,8 @@ cd ../pyexcel-presentation
 python setup.py install
 cd ../../
 nosetests --rednose --with-cov --with-doctest --doctest-extension=.rst doc/source pyexcel tests
-rm tmp.db
+if [ $? == 0 ] ; then
+	rm tmp.db
+else
+    exit 1;
+fi
